before reply page

+++ src/main/java/com/codeqa/forum/controller/PublishController.java
@@ -1,7 +1,6 @@
 package com.codeqa.forum.controller;
 
 import com.codeqa.forum.dto.QuestionDTO;
-import com.codeqa.forum.mapper.QuestionMapper;
 import com.codeqa.forum.model.Question;
 import com.codeqa.forum.model.User;
 import com.codeqa.forum.service.QuestionService;
@@ -27,7 +26,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model) {
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
@@ -42,7 +41,7 @@ public class PublishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "id", required = false) Integer id,
+            @RequestParam(value = "id", required = false) Long id,
             HttpServletRequest request,
             Model model) {
         model.addAttribute("title", title);
